+++ pkg/deviceclaimingserver/enddevices/enddevices.go
@@ -24,7 +24,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/auth/rights"
 	"go.thethings.network/lorawan-stack/v3/pkg/cluster"
 	"go.thethings.network/lorawan-stack/v3/pkg/config"
-	"go.thethings.network/lorawan-stack/v3/pkg/deviceclaimingserver/enddevices/ttjs"
+	"go.thethings.network/lorawan-stack/v3/pkg/deviceclaimingserver/enddevices/ttjsv2"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/fetch"
 	"go.thethings.network/lorawan-stack/v3/pkg/httpclient"
@@ -56,9 +56,9 @@ type Component interface {
 	AllowInsecureForCredentials() bool
 }
 
-const ttJSType = "ttjs"
-
-var errInvalidUpstream = errors.DefineInvalidArgument("invalid_upstream", "upstream `{type}` is invalid")
+const (
+	ttjsV2Type = "ttjsv2"
+)
 
 // Upstream abstracts EndDeviceClaimingServer.
 type Upstream struct {
@@ -91,41 +91,38 @@ func NewUpstream(ctx context.Context, conf Config, c Component, opts ...Option) 
 
 	// Setup upstreams.
 	for _, js := range baseConfig.JoinServers {
-		var (
-			s          EndDeviceClaimer
-			clientName string
-		)
-		switch js.Type {
-		case ttJSType:
-			// Fetch and parse configuration.
-			fileParts := strings.Split(filepath.ToSlash(js.File), "/")
-			fetcher := fetch.WithBasePath(fetcher, fileParts[:len(fileParts)-1]...)
-			fileName := fileParts[len(fileParts)-1]
-			configBytes, err := fetcher.File(fileName)
-			if err != nil {
-				return nil, err
-			}
-
-			var ttjsConfig ttjs.Config
-			if err := yaml.UnmarshalStrict(configBytes, &ttjsConfig); err != nil {
-				return nil, err
-			}
-
-			ttjsConfig.NetID = conf.NetID
-			ttjsConfig.JoinEUIPrefixes = js.JoinEUIs
-			ttjsConfig.NetworkServer.HomeNSID = conf.NetworkServer.HomeNSID
-			ttjsConfig.NetworkServer.Hostname = conf.NetworkServer.Hostname
-
-			s, err = ttjsConfig.NewClient(ctx, c)
-			if err != nil {
-				return nil, err
-			}
-			// The file for each client will be unique.
-			clientName = strings.Trim(fileName, filepath.Ext(fileName))
-		default:
-			return nil, errInvalidUpstream.WithAttributes("type", js.Type)
+		// Fetch and parse configuration.
+		fileParts := strings.Split(filepath.ToSlash(js.File), "/")
+		fetcher := fetch.WithBasePath(fetcher, fileParts[:len(fileParts)-1]...)
+		fileName := fileParts[len(fileParts)-1]
+		configBytes, err := fetcher.File(fileName)
+		if err != nil {
+			return nil, err
 		}
-		upstream.servers[clientName] = s
+
+		var claimer EndDeviceClaimer
+		switch js.Type {
+		case ttjsV2Type:
+			var cfg ttjsv2.Config
+			if err := yaml.UnmarshalStrict(configBytes, &cfg); err != nil {
+				return nil, err
+			}
+			cfg.NetID = conf.NetID
+			cfg.JoinEUIPrefixes = js.JoinEUIs
+			cfg.NetworkServer.Hostname = conf.NetworkServer.Hostname
+			cfg.NetworkServer.HomeNSID = conf.NetworkServer.HomeNSID
+			claimer, err = cfg.NewClient(ctx, c)
+			if err != nil {
+				return nil, err
+			}
+		default:
+			log.FromContext(ctx).WithField("type", js.Type).Warn("Unknown Join Server type")
+			continue
+		}
+
+		// The file for each client will be unique.
+		clientName := strings.Trim(fileName, filepath.Ext(fileName))
+		upstream.servers[clientName] = claimer
 	}
 
 	for _, opt := range opts {
@@ -151,21 +148,18 @@ var (
 )
 
 func (upstream *Upstream) joinEUIClaimer(ctx context.Context, joinEUI types.EUI64) EndDeviceClaimer {
-	for name, srv := range upstream.servers {
-		if !srv.SupportsJoinEUI(joinEUI) {
-			continue
+	for _, srv := range upstream.servers {
+		if srv.SupportsJoinEUI(joinEUI) {
+			return srv
 		}
-		log.FromContext(ctx).WithFields(log.Fields(
-			"name", name,
-			"join_eui", joinEUI,
-		)).Debug("JoinEUI supported by upstream")
-		return srv
 	}
 	return nil
 }
 
 // Claim implements EndDeviceClaimingServer.
-func (upstream *Upstream) Claim(ctx context.Context, joinEUI, devEUI types.EUI64, claimAuthenticationCode string) error {
+func (upstream *Upstream) Claim(
+	ctx context.Context, joinEUI, devEUI types.EUI64, claimAuthenticationCode string,
+) error {
 	claimer := upstream.joinEUIClaimer(ctx, joinEUI)
 	if claimer == nil {
 		return errClaimingNotSupported.WithAttributes("eui", joinEUI)
@@ -198,7 +192,9 @@ func (upstream *Upstream) Unclaim(ctx context.Context, in *ttnpb.EndDeviceIdenti
 }
 
 // GetInfoByJoinEUI implements EndDeviceClaimingServer.
-func (upstream *Upstream) GetInfoByJoinEUI(ctx context.Context, in *ttnpb.GetInfoByJoinEUIRequest) (*ttnpb.GetInfoByJoinEUIResponse, error) {
+func (upstream *Upstream) GetInfoByJoinEUI(
+	ctx context.Context, in *ttnpb.GetInfoByJoinEUIRequest,
+) (*ttnpb.GetInfoByJoinEUIResponse, error) {
 	joinEUI := types.MustEUI64(in.JoinEui).OrZero()
 	claimer := upstream.joinEUIClaimer(ctx, joinEUI)
 	return &ttnpb.GetInfoByJoinEUIResponse{
@@ -208,7 +204,9 @@ func (upstream *Upstream) GetInfoByJoinEUI(ctx context.Context, in *ttnpb.GetInf
 }
 
 // GetClaimStatus implements EndDeviceClaimingServer.
-func (upstream *Upstream) GetClaimStatus(ctx context.Context, in *ttnpb.EndDeviceIdentifiers) (*ttnpb.GetClaimStatusResponse, error) {
+func (upstream *Upstream) GetClaimStatus(
+	ctx context.Context, in *ttnpb.EndDeviceIdentifiers,
+) (*ttnpb.GetClaimStatusResponse, error) {
 	if in.DevEui == nil || in.JoinEui == nil {
 		return nil, errNoEUI.New()
 	}
@@ -227,7 +225,9 @@ func (upstream *Upstream) GetClaimStatus(ctx context.Context, in *ttnpb.EndDevic
 	return claimer.GetClaimStatus(ctx, in)
 }
 
-func (upstream *Upstream) requireRights(ctx context.Context, in *ttnpb.EndDeviceIdentifiers, appRights *ttnpb.Rights) error {
+func (upstream *Upstream) requireRights(
+	ctx context.Context, in *ttnpb.EndDeviceIdentifiers, appRights *ttnpb.Rights,
+) error {
 	// Collaborator must have the required rights on the application.
 	if err := rights.RequireApplication(ctx, in.ApplicationIds,
 		appRights.Rights...,